Move isUniqueFor functionality in the uniqueSlug file
