+++ backend/src/middleware/slugify/uniqueSlug.js
@@ -13,3 +13,17 @@ export default async function uniqueSlug (string, isUnique) {
   } while (!await isUnique(uniqueSlug))
   return uniqueSlug
 }
+
+export function isUniqueFor(context, type) {
+  return async slug => {
+    const session = context.driver.session()
+    const response = await session.run(
+      `MATCH(p:${type} {slug: $slug }) return p.slug`,
+      {
+        slug
+      }
+    )
+    session.close()
+    return response.records.length === 0
+  }
+}
